BUG 6490: Execute the APP_CACHE_VIEW triggers after we have subprocess data.
When creating a subprocess, a new row is added to the APP_DELEGATION table,
which triggers the insert triggers and updates APP_CACHE_VIEW. However, after
derivating, new data is added to the new case and then the APP_DELEGATION was
not being updated. This fix it by adding one second to the init date so that
an update is forced in APP_DELEGATION which executes the triggers and updates
APP_CACHE_VIEW with the right data.