BUG 6987 Error en PM al guardar formulario [FIXED]
When dropdown is with savelabel active then on form save the sql(s) need to be executed. By default any sqlconnection is defined as an empty string and Pm uses the default connection workflow.. the issue was that in savelabel stage this default option was not defined then tried to connect to an empty connection.