BUG 0000 Addition of the path to the plugin in the method uninstallPlugin()
We performed the inclusion of the file php with require_once, for handling it.
Then took place what is necessary for to uninstall the plugin.

This method is not being used anywhere in the core of ProcessMaker. We conducted a search with the following command:

find -name '*.*' -type f | xargs grep 'uninstallPlugin'