BUG 9180 Correct display priority in cases lists
Code Issue
----------
A number is display in the Priority column in the cases lists instead of the
priority in text.

Solution
--------
Correct display of priority column in cases lists.
The text that describes the priority is display instead of the priority number.
The number is converted to the corresponding text with the help of the function
G::LoadTranslation("ID_PRIORITY_N");