BUG-9449 Error in reassign and review of supervisor
Error

Error in reassign and review of supervisor. This is caused because Solr
don't support those views.

Solution

The filter to use Solr or DB was corrected to use DB in the case of Supervisor views.

Also a new exception was created to inform that APP_DATA could not be unserialized,
and the logic is not going to stop when an error of unserialization is run in the reindex process.