Bug 8626 Backup issues SOLVED
PROBLEM the backup is too bigger to be handle in some os configurations systema files
SOLUTION A new class will help to compress into several files instead one.
to use the new feature the admin would use and extra option as follow:
processmaker workspace-backup -s {}
to restore may use :
processmaker workspace-restore -m
In both case there are to new arguments -s and -m, these are for use the new feature only, if they are not added the old way is still available.