BUG 9873 Error creating Dynaform from a PM Table.
When adding a field, the 'Field Name' must be an string without spaces in the middle, because this makes the name of some funtions were generated wrong, like getColosa Email() when it must be like getColosaEmail() or getColosa_Email(). So, this field in the creating PM Table form, was validated to transform any string into a string without spaces.