BUG 10953Report Table error with currency fields
When having a process with some currency field (Real Number), and running a case before creating the RT, when you create the RT and saw the data inside it shows the number as integer. But when you run a case after creating the RT it shows the currency in good way (Real number).

It was fixed.