BUG 11246 "P.M.2.0.46Testing11: Department's supervisor's flaws" SOLVED
-When you set a user as supervisor in: Admin/User/Departaments, if i
made a mistake and i dont want any user to be the departments
supervisor, there should be an Unset Supervisor option. When you add
users to the department, the last user that its added is set up as
supervisor.
-Updgraded, according to the modules original design, there should
always be a departments supervisor, thats why all the asked options are
not being added. It has been made an upgrade to the clauses so there
should always be a departments supervisor.