BUG-12724 Issue with the task nottification
Descripcion: When you change the task name the task notification es disabled.

Solucion: It was resolved getting the value of the Task Notification and if it don´t  have any value assigned it will be assined FALSE value (disabled).
The issue with the Assigment tab was resolved and it's related with the http://bugs.processmaker.com/view.php?id=12467.
It was resolved for PM 2.5.2