BUG-14945 Disable foreign keys verification when generating a backup.
The statement SET FOREIGN_KEY_CHECKS=0; was added when the sql script is been generated, when backuping a workspace. This was added in case that a table in the database has a foreign key.

When a table has a foreign key some operations are not allowed create/delete/update, because of the restrictions of the foreing key. So, to allow this operations we disabled the foreign key checks (SET FOREIGN_KEY_CHECKS=0;) until we have all the tables created in the database.