BUG-15068 PMFPauseCase dosen`t pause the case whrn using the trigger wizzard.
The validation for the not required parameter "unpauseDate" was wrong. When this parameter was empty the validation was taking this like this has a date, so it was trying to check the date format, and this was giving an error.

The validation was changed to only let strings of 10 or more characters go into the next validation where the format of the date is checked.