BUG-13874 Duplicate typing for dynaforms over IE 11
When a dynaform has some fields with validations, when typing some data was duplicated.

I added a validation to detect IE version, if this is 11 a meta tag is added in the HTML (<meta http-equiv="X-UA-Compatible" content="IE=8"/>). This tag let the browser use the IE8 compatibility and prevent the duplicated characters issue.

The validation for IE 11 is a preg_match php function searching a 'Trident/' string. This is diferent in IE10, where the User-agent has MSIE on it. On IE 11 MSIE was deleted and the only way to detect that the browser is IE11 is searching 'Trident/'.