BUG-15116 InputFile is lost when importing a process.
When an already created process is imported with the 'Create a completely new process..' option, the inputs asociated in the dynaforms are not the input created for this new process, this new (copy) process still has associated the input of the original process (same UID), this because the uid of the new input is never replaced in the new process's dynaform xml file.

So it was created a new field in the data object (the data that has all the information about the new process) where is all the old inputs uid and the new ones. Then using this object this old uids are replaced with the new inputs uids.