BUG-15193 Error in PM Table Data Search option
The search option in PMTables has a validation where the APP_UID field is ignored in the search because this was causing some search troubles when you wanna search a number that was include into the APP_UID (eg. 'APP_UID':405044408538cb9472ade260172, 'text to search':4408 )

Now, we add a checkbox field near the search text field to customize this behavior. With this checkbox 'Search also in the APP_UID field' you can decide if the search will include or not the APP_UID field.