BUG-15384 Issue in the field check-box from Dynaform with the label.
When a label in a required field has a line break, a js error happens.

When the eval function is applied to the label's string, the line break is taked as the conclusion of the string, and js searchs " (double quotes) which closes the string without finding it and giving the error: "SintaxError: unterminated string literal"

So I use the replace function to change the break line to a simple space, this solves the problem and the error does not happend again.